setup: Setup started code for client and proxy
For the client code, we have a create-react-app typescript starter application along with a development Dockerfile. The proxy is a basic nginx image with a basic custom Dockerfile and config. Docker-compose is set up to work out of the box for convenient development. App runs on port 8080 when run with docker-compose.